requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
anthropic==0.7.8
requests==2.31.0
psycopg2-binary==2.9.9
scikit-learn==1.3.2
scipy==1.11.4
python-dotenv==1.0.0
finnhub-python==2.4.20